requirements.txt
Django~=4.2.6
djangorestframework~=3.14.0
django-filter~=23.3
drf-api-logger~=1.1.14
Markdown~=3.5
PyYAML~=6.0.1
uritemplate~=4.1.1
